// apps/admin/src/lib/api/generated/models/graphNode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.1
 */
import type { GraphNodeKind } from './graphNodeKind';

export interface GraphNode {
  description: string | null;
  entity_id: string;
  id: string;
  kind: GraphNodeKind;
  label: string;
  status: string | null;
  subtype: string | null;
}